Move leaf_function tests into test_leaf_function.py (4/4)
[ghstack-poisoned]
